change predict price to double, inheritance fund record

+++ src/hui_management_backend.Web/Endpoints/FundEndpoints/FundAddSession.RequestBody.cs
@@ -3,5 +3,5 @@
 public class FundAddSessionRequestBody
 {
   public int memberId { get; set; }
-  public int predictPrice { get; set; }
+  public double predictPrice { get; set; }
 }
